monumental update #2
the first huge update was when we started using arrays. this update includes an entire re-write for detecting when objects exist in a row, then writing the row with ordered objects. we used to use several temp arrays, sort, calculate distance, store distances in an array, ect. now we just create a row with empty spaces and overwrite spaces with objects using the object's X position.