middle of changing apple spawn, + other stuff
removed numOfTails function. just a variable now

removed setTails function. I haven't been using it during testing

in the middle of writing a new way of checking if the apple spawn loc is empty or not. should be much better. just need to handle what I'll do when I try spawning in a loc that already contains an obj